implement expression tree for complex query and Paged row response db query

+++ Backend/Application/Features/Common/Models/PagedResult.cs
@@ -7,7 +7,7 @@ namespace Application.Features.Common.Models;
 public class PagedResult<T>
 {
     public List<T> Items { get; set; } = new();
-    public int Total { get; set; }
+    public long Total { get; set; }
     public int Page { get; set; }
     public int PageSize { get; set; }
 }
